+++ pkg/registry/etcdregistry.go
