+++ pkg/registry/etcdregistry_test.go
